Folds the fully connected layer into the base class
